Moved ServiceTaskScheduler.java
