+++ src/main/java/xyz/rtsvk/alfax/services/ServiceManager.java
@@ -1,7 +1,6 @@
 package xyz.rtsvk.alfax.services;
 
 import xyz.rtsvk.alfax.util.Logger;
-import xyz.rtsvk.alfax.util.ServiceTaskScheduler;
 import xyz.rtsvk.alfax.util.text.TextUtils;
 
 import java.util.HashMap;
